Update subversion plugin to work with Subversion v1.7.
Subversion 1.7 has a single .svn directory in the root of the respository unlike
earlier versions which kept meta data in every versioned directory. The change
means we can no longer tell if we're in a subversion controlled directory by
looking for that folder.

The fix is to try and use the svn info command instead and see if it returns an
error.

An alternative would have been to use svn status which will also tell us if
we're in a subdirectory that hasn't been added to svn but is a child directory
of an svn controlled repository. However, in this scenario it becomes difficult
to extract information about the svn repository as svn info won't work. A work
around would be to recurse through the parent directories until we found one
which did work but this might become to unperformant so for now I went with svn
info.

 Please enter the commit message for your changes. Lines starting